+++ src/Config/ConfigProvider.php
@@ -88,6 +88,6 @@ class ConfigProvider
      */
     public function getDefaultConfiguration(): array
     {
-        return require dirname(__DIR__, 2) . '/config/soluble-mediatools.global.php';
+        return require dirname(__DIR__, 2) . '/config/soluble-mediatools.config.php';
     }
 }